chore: consolidate quantifier error tests into tests/errors/quantifier.test.py
